final tidy of useStore
